Move exchange validations to domain layer
