+++ src/datasources/exchange-api/exchange-api.service.ts
@@ -1,8 +1,4 @@
-import {
-  Inject,
-  Injectable,
-  InternalServerErrorException,
-} from '@nestjs/common';
+import { Inject, Injectable } from '@nestjs/common';
 import { ExchangeResult } from '../../domain/exchange/entities/exchange-result.entity';
 import {
   INetworkService,
@@ -11,11 +7,6 @@ import {
 import { IConfigurationService } from '../../common/config/configuration.service.interface';
 import { FiatCodesExchangeResult } from '../../domain/exchange/entities/fiat-codes-result.entity';
 import { IExchangeApi } from '../../domain/interfaces/exchange-api.interface';
-import { DefinedError, ValidateFunction } from 'ajv';
-import { ValidationErrorFactory } from '../../domain/errors/validation-error-factory';
-import { fiatCodesExchangeResultSchema } from './entities/schemas/fiat-codes-exchange-result.schema';
-import { JsonSchemaService } from '../../common/schemas/json-schema.service';
-import { exchangeResultSchema } from './entities/schemas/exchange-result.schema';
 
 @Injectable()
 export class ExchangeApi implements IExchangeApi {
@@ -23,79 +14,30 @@ export class ExchangeApi implements IExchangeApi {
 
   private readonly baseUrl: string;
   private readonly apiKey: string;
-  private readonly isValidExchangeResult: ValidateFunction<ExchangeResult>;
-  private readonly isValidFiatCodesExchangeResult: ValidateFunction<FiatCodesExchangeResult>;
 
   constructor(
     @Inject(IConfigurationService)
     private readonly configurationService: IConfigurationService,
     @Inject(NetworkService) private readonly networkService: INetworkService,
-    private readonly validationErrorFactory: ValidationErrorFactory,
-    private readonly jsonSchemaService: JsonSchemaService,
   ) {
     this.baseUrl =
       this.configurationService.getOrThrow<string>('exchange.baseUri');
     this.apiKey =
       this.configurationService.getOrThrow<string>('exchange.apiKey');
-
-    this.isValidExchangeResult = this.jsonSchemaService.compile(
-      exchangeResultSchema,
-    ) as ValidateFunction<ExchangeResult>;
-
-    this.isValidFiatCodesExchangeResult = this.jsonSchemaService.compile(
-      fiatCodesExchangeResultSchema,
-    ) as ValidateFunction<FiatCodesExchangeResult>;
   }
 
-  async convertRates(to: string, from: string): Promise<number> {
-    const exchangeResult = await this.getExchangeResult();
-
-    if (exchangeResult.rates === undefined)
-      throw new InternalServerErrorException(`Exchange rates unavailable`);
-
-    const fromExchangeRate = exchangeResult.rates[from.toUpperCase()];
-    if (fromExchangeRate === undefined || fromExchangeRate == 0)
-      throw new InternalServerErrorException(
-        `Exchange rate for ${from} is not available`,
-      );
-    const toExchangeRate = exchangeResult.rates[to.toUpperCase()];
-    if (toExchangeRate === undefined)
-      throw new InternalServerErrorException(
-        `Exchange rate for ${to} is not available`,
-      );
-
-    return toExchangeRate / fromExchangeRate;
-  }
-
-  async getFiatCodes(): Promise<string[]> {
-    const fiatCodesResult = await this.getFiatCodesExchangeResult();
-
-    return Object.keys(fiatCodesResult.symbols);
-  }
-
-  private async getExchangeResult(): Promise<ExchangeResult> {
-    const { data } = await this.networkService.get(`${this.baseUrl}/latest`, {
-      params: { access_key: this.apiKey },
-    });
-
-    if (!this.isValidExchangeResult(data)) {
-      const errors = this.isValidExchangeResult.errors as DefinedError[];
-      throw this.validationErrorFactory.from(errors);
-    }
-
-    return data;
-  }
-
-  private async getFiatCodesExchangeResult(): Promise<FiatCodesExchangeResult> {
+  async getFiatCodes(): Promise<FiatCodesExchangeResult> {
     const { data } = await this.networkService.get(`${this.baseUrl}/symbols`, {
       params: { access_key: this.apiKey },
     });
 
-    if (!this.isValidFiatCodesExchangeResult(data)) {
-      const errors = this.isValidFiatCodesExchangeResult
-        .errors as DefinedError[];
-      throw this.validationErrorFactory.from(errors);
-    }
+    return data;
+  }
+
+  async getExchangeResult(): Promise<ExchangeResult> {
+    const { data } = await this.networkService.get(`${this.baseUrl}/latest`, {
+      params: { access_key: this.apiKey },
+    });
 
     return data;
   }
