+++ src/domain/exchange/entities/exchange-result.entity.ts
@@ -1,6 +1,6 @@
 export interface ExchangeResult {
   success: boolean;
   // TODO: use Record<string, number>, define a compatible JSONSchema
-  rates?: object;
+  rates: object;
   base: string;
 }
